Use ArchivePreparer instead of SecuredImpl for getting and adding keycloak.json
