+++ fractions/keycloak/src/main/java/org/wildfly/swarm/keycloak/internal/SecuredImpl.java
@@ -15,22 +15,10 @@
  */
 package org.wildfly.swarm.keycloak.internal;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-
-import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.Node;
-import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.Asset;
-import org.jboss.shrinkwrap.api.asset.ByteArrayAsset;
-import org.jboss.shrinkwrap.api.importer.ZipImporter;
 import org.jboss.shrinkwrap.impl.base.ArchiveBase;
 import org.jboss.shrinkwrap.impl.base.AssignableBase;
-import org.wildfly.swarm.bootstrap.util.BootstrapProperties;
 import org.wildfly.swarm.keycloak.Secured;
 import org.wildfly.swarm.spi.api.JARArchive;
 import org.wildfly.swarm.undertow.descriptors.SecurityConstraint;
@@ -67,23 +55,9 @@ public class SecuredImpl extends AssignableBase<ArchiveBase<?>> implements Secur
         getArchive().as(JARArchive.class).addModule("org.wildfly.swarm.keycloak", "deployment");
         getArchive().as(JARArchive.class).addAsServiceProvider("io.undertow.servlet.ServletExtension", "org.wildfly.swarm.keycloak.deployment.SecurityContextServletExtension");
 
-        InputStream keycloakJson;
-        String keycloakJsonPath = System.getProperty("swarm.keycloak.json.path");
-        if (keycloakJsonPath != null) {
-            keycloakJson = getKeycloakJson(keycloakJsonPath);
-        } else {
-            keycloakJson = getKeycloakJson();
-        }
-
         // Setup web.xml
         this.asset.setContextParam("resteasy.scan", "true");
         this.asset.setLoginConfig("KEYCLOAK", "ignored");
-
-        if (keycloakJson != null) {
-            getArchive().as(JARArchive.class).add(createAsset(keycloakJson), "WEB-INF/keycloak.json");
-        } else {
-            // not adding it.
-        }
     }
 
     @Override
@@ -96,59 +70,7 @@ public class SecuredImpl extends AssignableBase<ArchiveBase<?>> implements Secur
         return this.asset.protect(urlPattern);
     }
 
-    private Asset createAsset(InputStream in) {
-
-        StringBuilder str = new StringBuilder();
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(in))) {
-
-            String line = null;
-
-            while ((line = reader.readLine()) != null) {
-                str.append(line).append("\n");
-            }
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return new ByteArrayAsset(str.toString().getBytes());
-    }
-
     private WebXmlAsset asset;
-
-    private InputStream getKeycloakJson(String path) {
-        try {
-            return Files.newInputStream(Paths.get(path));
-        } catch (IOException ignore) {}
-
-        return null;
-    }
-
-    private InputStream getKeycloakJson() {
-        InputStream keycloakJson = Thread.currentThread().getContextClassLoader().getResourceAsStream("keycloak.json");
-        if (keycloakJson == null) {
-
-            String appArtifact = System.getProperty(BootstrapProperties.APP_ARTIFACT);
-
-            if (appArtifact != null) {
-                try (InputStream in = ClassLoader.getSystemClassLoader().getResourceAsStream("_bootstrap/" + appArtifact)) {
-                    Archive tmpArchive = ShrinkWrap.create(JARArchive.class);
-                    tmpArchive.as(ZipImporter.class).importFrom(in);
-                    Node jsonNode = tmpArchive.get("keycloak.json");
-                    if (jsonNode == null) {
-                        jsonNode = tmpArchive.get("WEB-INF/keycloak.json");
-                    }
-
-                    if (jsonNode != null && jsonNode.getAsset() != null) {
-                        keycloakJson = jsonNode.getAsset().openStream();
-                    }
-                } catch (IOException e) {
-                    // ignore
-                    // e.printStackTrace();
-                }
-            }
-        }
-        return keycloakJson;
-    }
 
 }
